composer require php-di/php-di -W
-W was required to get dependencies.  Otherwise I got this error:

Your requirements could not be resolved to an installable set of packages.

  Problem 1
    - php-di/php-di[6.3.0, ..., 6.3.5] require psr/container ^1.0 -> found psr/container[1.0.0, 1.1.0, 1.1.1] but the pa
ckage is fixed to 2.0.1 (lock file version) by a partial update and that version does not match. Make sure you list it a
s an argument for the update command.
    - Root composer.json requires php-di/php-di ^6.3 -> satisfiable by php-di/php-di[6.3.0, ..., 6.3.5].

Use the option --with-all-dependencies (-W) to allow upgrades, downgrades and removals for packages currently locked to
specific versions.